Require nasapower >= 4.0.0
Since old versions of _naspower_ no longer work at all it's probably best to enforce v4.0.0 and above for compatibility with the POWER API when installing _apsimx_. I just received an e-mail from someone using your package getting an error 422, which is likely due to an out-of-date _naspower_ installation. That won't be fixed for prior installs of _apsimx_ but we can head it off with new installations.